+++ OnlineCourses.Domain.Layer/Model/CourseListDto.cs
@@ -6,6 +6,6 @@
         public int TotalSignatures { get; set; }
         public int MinimumAge { get; set; }
         public int MaximumAge { get; set; }
-        public int AverageAge { get; set; }
+        public double AverageAge { get; set; }
     }
 }
